requirements.txt
pandas==1.4.2
PyQt5==5.15.6
pyqtgraph==0.12.4
pyserial==3.5
SQLAlchemy==1.4.36
